% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{cal_prob}
\alias{cal_prob}
\title{Calibrate sample}
\usage{
cal_prob(data, cal, current, dwell_time, accumulations)
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{cal}{Calibration data from cal_std() function}

\item{current}{EMP current in µA used during the measurement session}

\item{dwell_time}{Spot dwell time}

\item{accumulations}{Number of accumulations}
}
\description{
This function helps to calibrate a sample using the output of the cal_std() function
}
