% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{cal_std}
\alias{cal_std}
\title{Prepare calibration data}
\usage{
cal_std(data, B, cal)
}
\arguments{
\item{data}{standard file import from get_std() function}

\item{B}{B parameter from get_bg() function. If B is not supplied, it will be estimated by the function.}

\item{cal}{known element concentrations for standards}
}
\description{
This function calculates the calibration parameters A and B.
Please refer to the vignette for more details.
}
\keyword{calibrate}
