% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{contrast}
\alias{contrast}
\title{Contrast colour palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
contrast
}
\description{
Contrast colour palette
}
\keyword{datasets}
