% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{cut_prob}
\alias{cut_prob}
\title{Threshold filter}
\usage{
cut_prob(data, type = c("fixed", "quant"), cut_threshold)
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{type}{Fixed or quantile threshold}

\item{cut_threshold}{Defined threshold. Must be numeric}
}
\description{
Remove all data points below a fixed or a quantile threshold.
This function is helpful to remove noise around sample edges.
Please refer to the vignette for more details.
}
