% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{dens_prob}
\alias{dens_prob}
\title{Plot density function}
\usage{
dens_prob(data, return_fig = FALSE)
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{return_fig}{Logical specifying if a ggplot figure should be returned by the function instead of a data frame. Default is FALSE.}
}
\description{
Plot density function
}
