% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{flip_prob}
\alias{flip_prob}
\title{Flip function}
\usage{
flip_prob(data, flip_dir = c("h", "v"))
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{flip_dir}{Flip horizontally 'h' or vertically 'v'}
}
\description{
Flip sample horizontally or vertically
}
