% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_bg}
\alias{get_bg}
\title{Import background}
\usage{
get_bg(data_path, data_string)
}
\arguments{
\item{data_path}{Path to file(s)}

\item{data_string}{File name(s).}
}
\description{
This function helps to import background files to EMPiR.
Data must be an n by m matrix without header or row ID stored in individual .txt files.
Please refer to the vignette for more details.
}
\examples{
get_bg('path/to/file', 'Bg') 
}
\keyword{background}
\keyword{get,}
\keyword{import,}
