% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_std}
\alias{get_std}
\title{Import standard files}
\usage{
get_std(data_path, data_string)
}
\arguments{
\item{data_path}{Path to file(s)}

\item{data_string}{File name(s)}
}
\description{
This function helps to import standard files to EMPiR.
Data must be an n by m matrix without header or row ID stored in individual .txt files.
Please refer to the vignette for more details.
}
\examples{
get_prob('path/to/file', c('VG-2', 'Dolomite')) 
}
\keyword{get,}
\keyword{import,}
\keyword{standard}
