% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{hist_prob}
\alias{hist_prob}
\title{Plot histogram}
\usage{
hist_prob(
  data,
 
    palette = c("contrast,\\n                                   moonshine,\\n                                   pastell"),
  bins = 50
)
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{palette}{Select color palette}

\item{bins}{Bin size used in histogram}
}
\description{
Plot histogram
}
