% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{image_prob}
\alias{image_prob}
\title{Plot raster}
\usage{
image_prob(data, legend_name, palette, scale_location, beam_size)
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{legend_name}{Legend title. Useful to indicate sample name, element and units.}

\item{palette}{Select colour palette.}

\item{scale_location}{Position of the scale bar}

\item{beam_size}{Size of the beam used to calculate true size from pixel.}
}
\description{
Main plot function.
}
