% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{moonlight}
\alias{moonlight}
\title{Moonlight colour palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
moonlight
}
\description{
Moonlight colour palette
}
\keyword{datasets}
