% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{smooth_prob}
\alias{smooth_prob}
\title{Smooth function}
\usage{
smooth_prob(data, fac)
}
\arguments{
\item{data}{Sample data from get_prob() function}

\item{fac}{Smoothing factor}
}
\description{
Applies a raster smoothing function to reduce noise.
Please refer to the vignette for more details.
}
