% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operate.R
\name{subset_rect}
\alias{subset_rect}
\title{Subset rectangle
Unfortunately R does not yet provide an option to manually subset ggplots by mouse select.
We thus define two points between which a subset rectangle will be created.
Input points are on a relative scale 0-100 rather than on an absolute
scale to make it easier for the operator to select the desired area.}
\usage{
subset_rect(data, x1, x2, y1, y2)
}
\arguments{
\item{data}{Sample data from get_prob() function}
}
\description{
Subset rectangle
Unfortunately R does not yet provide an option to manually subset ggplots by mouse select.
We thus define two points between which a subset rectangle will be created.
Input points are on a relative scale 0-100 rather than on an absolute
scale to make it easier for the operator to select the desired area.
}
