% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{cal_std}
\alias{cal_std}
\title{Prepare calibration data}
\usage{
cal_std(data, B, cal, current, dwell_time, accumulations)
}
\arguments{
\item{data}{standard file import from get_std() function}

\item{B}{B parameter from get_bg() function. If B is not supplied, it will be estimated by the function.}

\item{cal}{known element concentrations for standards}

\item{current}{EMP current in microampere used in the measurement session}

\item{dwell_time}{Spot dwell time}

\item{accumulations}{Number of accumulations}
}
\description{
This function calculates the calibration parameters A and B.
Please refer to the vignette for more details.
}
\examples{
cal_std(cal = c('VG-2' = 4.05, 
 'Dolomite' = 13.29),
 B = B,
 current=0.1004,
 dwell_time = 20,
 accumulations = 8)
}
\keyword{calibrate}
